Renamed mainAudio and mainAudioFromFile (#355)
